tests & fixes
